feat: move cms articles from labs
