+++ libs/domain/content/src/index.ts
@@ -1,3 +1,4 @@
 export * from './feature';
 export * from './models';
+export * from './resolvers';
 export * from './services';
