+++ libs/domain/content/src/models/article.model.ts
@@ -1,0 +1,9 @@
+// TODO: create dynamic types and dependable dynamic pages.
+export const articleTypes = ['article', 'faq', 'about'];
+
+export interface ArticleContent {
+  heading: string;
+  description: string;
+  content: string;
+  type: string;
+}
