+++ libs/domain/content/src/models/content.model.ts
@@ -1,6 +1,9 @@
 declare global {
   interface ContentFields {
     component: undefined;
+    article: undefined;
+    faq: undefined;
+    about: undefined;
   }
 }
