+++ libs/domain/content/src/models/index.ts
@@ -1,3 +1,4 @@
+export * from './article.model';
 export * from './content.model';
 
 /**
