+++ libs/domain/content/src/services/article-context.ts
@@ -1,8 +1,8 @@
-import { ContentContext } from '@oryx-frontend/content';
 import { ContextFallback } from '@oryx-frontend/core';
 import { inject, Provider } from '@oryx-frontend/di';
 import { RouterService } from '@oryx-frontend/router';
 import { combineLatest, map } from 'rxjs';
+import { ContentContext } from './content-context';
 
 export const ArticleQualifierContextFallback: Provider = {
   provide: `${ContextFallback}${ContentContext.Content}`,
