+++ libs/domain/content/src/services/content.providers.ts
@@ -1,7 +1,10 @@
-import { provideEntity } from '@oryx-frontend/core';
+import { PageMetaResolver, provideEntity } from '@oryx-frontend/core';
 import { Provider } from '@oryx-frontend/di';
 import { ExperienceAdapter } from '@oryx-frontend/experience';
-import { ContentExperienceAdapter } from './adapter';
+import { articleTypes, Content, ContentQualifier } from '../models';
+import { ArticlePageDescriptionMetaResolver, ArticlePageTitleMetaResolver } from '../resolvers';
+import { ContentConfig, ContentExperienceAdapter } from './adapter';
+import { ArticleQualifierContextFallback } from './article-context';
 import {
   contentfulProviders,
   storyblokProviders,
@@ -32,4 +35,36 @@ export const contentProviders: Provider[] = [
     service: ContentService,
     context: ContentContext.Content,
   }),
+  ArticleQualifierContextFallback,
+  {
+    provide: PageMetaResolver,
+    useClass: ArticlePageTitleMetaResolver,
+  },
+  {
+    provide: PageMetaResolver,
+    useClass: ArticlePageDescriptionMetaResolver,
+  },
+  {
+    provide: ContentConfig,
+    useValue: {
+      storyblok: {
+        types: ['component', 'faq', 'contents'],
+      },
+      strapi: {
+        types: ['component', 'about', 'contents'],
+        defaultType: 'about',
+      },
+      contentful: {
+        types: ['component', 'article', 'contents'],
+      },
+    },
+  },
+  ...articleTypes.map((type) =>
+    provideEntity<Content | null | undefined, ContentQualifier>(type, {
+      service: ContentService,
+      context: type,
+      get: (service, qualifier) =>
+        (service as ContentService).get({ ...qualifier, type }),
+    })
+  ),
 ];
