--- libs/template/labs/src/articles/article-types.ts
@@ -1,1 +1,0 @@
-export const articleTypes = ['article', 'faq', 'about'];
