--- libs/template/labs/src/articles/article.model.ts
@@ -1,14 +1,0 @@
-declare global {
-  interface ContentFields {
-    article: undefined;
-    about: undefined;
-    faq: undefined;
-  }
-}
-
-export interface ArticleContent {
-  heading: string;
-  description: string;
-  content: string;
-  type: string;
-}
