--- libs/template/labs/src/articles/index.ts
@@ -1,5 +1,0 @@
-export * from './article-context';
-export * from './article-page';
-export * from './article-providers';
-export * from './article.model';
-export * from './resolvers';
