+++ libs/template/labs/src/index.ts
@@ -1,2 +1,1 @@
-export * from './articles';
 export * from './feature';
