+++ libs/template/presets/storefront/experience/pages/index.ts
@@ -1,3 +1,4 @@
+export * from './article-page';
 export * from './cart-page';
 export * from './category-page';
 export * from './checkout-page';
@@ -9,3 +10,4 @@ export * from './order-confirmation-page';
 export * from './product-page';
 export * from './registration-page';
 export * from './search-page';
+
